+++ src/pages/CollapsePage.js
@@ -16,26 +16,17 @@ const names = [
 
 function Example() {
     return (
-        <div>
-            <Collapse defaultState={true} border>
-                <p className="px-10 fz-9 fw-bold">Some Article</p>
-                <p className="pa-10">
-                    Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
-                    tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
-                    quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
-                    consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
-                    cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non
-                    proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-                </p>
-            </Collapse>
-            <br/>
+        <>
             <Collapse 
                 title={<span className="mx-10 fz-10 fw-bold">Marvel Avengers</span>}
-                defaultState={true} 
+                defaultState 
                 border 
+                onChange={() => console.log('Do smth...')}
+                dark={theme}
                 tooltip="Expand/Collapse">
                 <List 
                     size="medium"
+                    className="py-10"
                     hover>
                     {names.map((item, index) => 
                         <List.Item 
@@ -46,7 +37,7 @@ function Example() {
                     )}
                 </List>
             </Collapse>
-        </div>
+        </>
     )
 }`
 const items = [
@@ -56,6 +47,13 @@ const items = [
         default: 'false', 
         type: 'boolean',
         value: 'true | false'
+    },
+    { 
+        property: 'onChange',
+        description: 'Invokes on collapse toggling (return current state)', 
+        default: '', 
+        type: 'function',
+        value: ''
     },
     { 
         property: 'border',
@@ -69,6 +67,13 @@ const items = [
         description: 'Set collapse title', 
         default: '', 
         type: 'string',
+        value: ''
+    },
+    { 
+        property: 'customToggler',
+        description: 'Set custom toggler element', 
+        default: '', 
+        type: 'ReactNode',
         value: ''
     },
     { 
@@ -137,26 +142,16 @@ const CollapsePage = () => {
                     usage={usage}
                     apiDescItems={items}
                     componentName="<Collapse/>">
-                    <Collapse dark={theme} defaultState border>
-                        <p className="px-10 fz-9 fw-bold">Some Article</p>
-                        <p className="pa-10">
-                            Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
-                            tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
-                            quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
-                            consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
-                            cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non
-                            proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-                        </p>
-                    </Collapse>
-                    <br/>
                     <Collapse 
                         title={<span className="mx-10 fz-10 fw-bold">Marvel Avengers</span>}
                         defaultState 
                         border 
+                        onChange={() => console.log('Do smth...')}
                         dark={theme}
                         tooltip="Expand/Collapse">
                         <List 
                             size="medium"
+                            className="py-10"
                             hover>
                             {names.map((item, index) => 
                                 <List.Item